+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - File name patterns: %F is the name of the source file without extension, but
   with the path.
 - A hint for pip installations without using `--no-compile`.
+- KiBoM variants: support to field overwrite according to variant (experimental).
 ### Fixed
 - Now we support missing field names in schematic library entries.